Move pickers into own subfolder.
