+++ src/components/pickers/EnumPicker.js
@@ -2,9 +2,9 @@
 import React, { useState } from 'react';
 
 /**
- * @typedef {import("../model/enums/EnumType").default} EnumType
- * @typedef {import("../typedefs").MapSetter} MapSetter
- * @typedef {import("../controller/requests/FilterRequest").FilterItem} FilterItem
+ * @typedef {import("../../model/enums/EnumType").default} EnumType
+ * @typedef {import("../../controller/requests/FilterRequest").FilterItem} FilterItem
+ * @typedef {import("../../typedefs").MapSetter} MapSetter
  */
 
  /**
